requirements.txt
albumentations==1.3.0
boto3==1.25.2
deepchecks[vision]==0.9.2
labelbox[data]==3.27.2
onnx==1.12.0
onnxruntime==1.13.1
opencv-python==4.6.0.66
pascal-voc==0.0.3
pillow==9.2.0
torchmetrics==0.10.1
torchvision==0.13.0
zenml[server]==0.20.5
